requirements.txt
# Librerie di base
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Gestione dati finanziari
yfinance>=0.1.70
alpha_vantage>=2.3.1
ta>=0.7.0  # Analisi tecnica

# Machine Learning e Deep Learning
scikit-learn>=1.0.0
tensorflow>=2.8.0
keras>=2.8.0
torch>=1.10.0
transformers>=4.18.0

# Web e API
flask>=2.0.0
flask-cors>=3.0.10
requests>=2.27.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
plotly>=5.13.0
dash>=2.9.0
dash-bootstrap-components>=1.3.0
dash-core-components>=2.0.0
dash-html-components>=2.0.0
dash-table>=5.0.0
pandas-datareader>=0.10.0
websockets>=11.0.3

# Utility
python-dotenv>=0.20.0
cryptography>=36.0.0
redis>=4.2.0
prometheus-client>=0.14.0
psutil>=5.9.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Development
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
isort>=5.10.0

# Caching and Performance
redis>=4.0.0
prometheus-client>=0.14.0
psutil>=5.9.0

# Security
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0

# Additional dependencies
matplotlib>=3.7.0
cryptography>=41.0.0

# Updated dependencies
yfinance>=0.2.18
ta>=0.10.2
scikit-learn>=1.2.2
tensorflow>=2.12.0
numpy>=1.24.2
pandas>=1.5.3
matplotlib>=3.7.1
seaborn>=0.12.2
python-dotenv>=1.0.0
requests>=2.28.2 